stock_analysis_agent/requirements.txt
acres==0.5.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.6
camelot-py==1.0.9
certifi==2025.11.12
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.5
curl_cffi==0.13.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.129.2
fastjsonschema==2.21.2
filelock==3.25.0
fqdn==1.5.1
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_git==0.51.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
multitasking==0.0.12
narwhals==2.16.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
notebook_shim==0.2.4
numpy==2.4.2
openai==2.21.0
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathlib==1.0.1
pdfminer.six==20260107
peewee==4.0.1
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.5
prov==2.1.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==2.0.0
pyarrow==23.0.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdf==5.9.0
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
pyzmq==27.1.0
rdflib==7.6.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.29.0
scipy==1.17.1
Send2Trash==1.8.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.52.1
streamlit==1.54.0
tabulate==0.9.0
tenacity==9.1.4
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.3
traitlets==5.14.3
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.41.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
yfinance==1.2.0
